requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.0.2
django-extensions==3.2.3
django-filter==24.3
django-heroku==0.3.1
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
